+++ src/connectedk8s/azext_connectedk8s/_utils.py
@@ -767,7 +767,8 @@ def helm_install_release(resource_manager, chart_path, subscription_id, kubernet
                          resource_group_name, cluster_name, location, onboarding_tenant_id, http_proxy, https_proxy,
                          no_proxy, proxy_cert, private_key_pem, kube_config, kube_context, no_wait, values_file,
                          cloud_name, disable_auto_upgrade, enable_custom_locations, custom_locations_oid,
-                         helm_client_location, enable_private_link, arm_metadata, onboarding_timeout="1200",
+                         helm_client_location, enable_private_link, arm_metadata,
+                         onboarding_timeout=consts.DEFAULT_MAX_ONBOARDING_TIMEOUT_HELMVALUE_SECONDS,
                          container_log_path=None):
 
     cmd_helm_install = [helm_client_location, "upgrade", "--install", "azure-arc", chart_path,